+++ DBHesabdari_PG/Models/EP/CodingHesabdari/EpGroupTafsiliLevel3.cs
@@ -27,9 +27,9 @@ namespace DBHesabdari_PG.Models.EP.CodingHesabdari
         //[Required, MaxLength(50)]
         //public string Level2Name { get; set; }
         [Required]
-        public int StartCode { get; set; }
+        public long StartCode { get; set; }
         [Required]
-        public int EndCode { get; set; }
+        public long EndCode { get; set; }
         [Required]
         public bool IsActive { get; set; }
         [MaxLength(500)]
